# Social Media Platform/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Social Media Platform API
  version: 1.0.0
  description: REST API documentation for the Selamat social media platform.
servers:
  - url: http://localhost:8080/api
    description: Local development server
  - url: https://api.selamat.example.com/api
    description: Production server placeholder
...
paths:
...
  /users/{userId}/profile:
    get:
      tags: [Profiles]
      summary: Get a user's profile
      operationId: getProfile
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Profile returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Profiles]
      summary: Update the current user's profile
      operationId: updateProfile
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /posts:
    post:
      tags: [Posts]
      summary: Create a new post
      operationId: createPost
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          description: Post created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /posts/{postId}:
    get:
      tags: [Posts]
      summary: Get a post by ID
      operationId: getPostById
      parameters:
        - $ref: '#/components/parameters/PostId'
      responses:
        '200':
          description: Post returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Posts]
      summary: Update a post owned by the current user
      operationId: updatePost
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PostId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostRequest'
      responses:
        '200':
          description: Post updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Posts]
      summary: Delete a post owned by the current user
      operationId: deletePost
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PostId'
      responses:
        '204':
          description: Post deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /media:
    post:
      tags: [Media]
      summary: Upload media for a post or profile
      operationId: uploadMedia
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                altText:
                  type: string
                  maxLength: 255
      responses:
        '201':
          description: Media uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Media'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /users/{userId}/follow:
    get:
      tags: [Follows]
      summary: Check whether the current user follows this user
      operationId: getFollowStatus
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Follow status returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FollowStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Follows]
      summary: Follow or request to follow a user
      operationId: followUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '201':
          description: Follow created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Follow'
        '202':
          description: Follow request pending
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Follow'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Follows]
      summary: Unfollow a user
      operationId: unfollowUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: Unfollowed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /messages:
    get:
      tags: [Messages]
      summary: List message conversations for the current user
      operationId: listConversations
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Messages returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagePage'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Messages]
      summary: Send a direct message
      operationId: sendMessage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /messages/{messageId}:
    get:
      tags: [Messages]
      summary: Get a message by ID
      operationId: getMessageById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MessageId'
      responses:
        '200':
          description: Message returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /messages/conversations/{participantId}/read:
    patch:
      tags: [Messages]
      summary: Mark a conversation as read
      operationId: markConversationAsRead
      security:
        - bearerAuth: []
      parameters:
        - name: participantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Conversation marked as read
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /notifications:
    get:
      tags: [Notifications]
      summary: List notifications for the current user
      operationId: listNotifications
      security:
        - bearerAuth: []
      parameters:
        - name: unreadOnly
          in: query
          required: false
          schema:
            type: boolean
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Notifications returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /notifications/unread-count:
    get:
      tags: [Notifications]
      summary: Get unread notification count
      operationId: getUnreadCount
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Unread count returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationCountResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /notifications/read-all:
    patch:
      tags: [Notifications]
      summary: Mark all notifications as read
      operationId: markAllNotificationsRead
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Unread count returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationCountResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /notifications/{notificationId}/read:
    patch:
      tags: [Notifications]
      summary: Mark a notification as read
      operationId: markNotificationRead
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/NotificationId'
      responses:
        '200':
          description: Notification updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PostId:
      name: postId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    MessageId:
      name: messageId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    NotificationId:
      name: notificationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
        format: date-time
      description: ISO-8601 cursor returned from a previous paginated response.
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication is required or the token is invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: The authenticated user does not have permission
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    UnprocessableEntity:
      description: Request was valid JSON but failed business validation or moderation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Profile:
      type: object
      required: [id, userId]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        displayName:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
          format: uri
        coverImageUrl:
          type: string
          format: uri
        street:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        websiteUrl:
          type: string
          format: uri
    UpdateProfileRequest:
      type: object
      properties:
        displayName:
          type: string
          maxLength: 100
        bio:
          type: string
          maxLength: 500
        avatarUrl:
          type: string
          format: uri
        coverImageUrl:
          type: string
          format: uri
          maxLength: 500
        street:
          type: string
          maxLength: 255
        city:
          type: string
          maxLength: 120
        state:
          type: string
          maxLength: 120
        country:
          type: string
          maxLength: 120
        websiteUrl:
          type: string
          format: uri
          maxLength: 255
    CreatePostRequest:
      type: object
      required: [visibility]
      properties:
        content:
          type: string
          maxLength: 5000
        visibility:
          $ref: '#/components/schemas/PostVisibility'
        mediaIds:
          type: array
          items:
            type: string
            format: uuid
    UpdatePostRequest:
      type: object
      properties:
        content:
          type: string
          maxLength: 5000
        visibility:
          $ref: '#/components/schemas/PostVisibility'
    Post:
      type: object
      required: [id, author, visibility, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        author:
          $ref: '#/components/schemas/UserSummary'
        content:
          type: string
        visibility:
          $ref: '#/components/schemas/PostVisibility'
        status:
          $ref: '#/components/schemas/PostStatus'
        media:
          type: array
          items:
            $ref: '#/components/schemas/Media'
        commentCount:
          type: integer
          format: int64
        reactionCount:
          type: integer
          format: int64
        viewerReaction:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/ReactionType'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Media:
      type: object
      required: [id, url, mediaType, mimeType, sizeBytes, createdAt]
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        mediaType:
          $ref: '#/components/schemas/MediaType'
        mimeType:
          type: string
          example: image/png
        sizeBytes:
          type: integer
          format: int64
        altText:
          type: string
        createdAt:
          type: string
          format: date-time
...
    Follow:
      type: object
      required: [id, followerId, followedId, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        followerId:
          type: string
          format: uuid
        followedId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/FollowStatus'
        createdAt:
          type: string
          format: date-time
    FollowStatusResponse:
      type: object
      required: [following]
      properties:
        following:
          type: boolean
    SendMessageRequest:
      type: object
      required: [recipientId, content]
      properties:
        recipientId:
          type: string
          format: uuid
        content:
          type: string
          maxLength: 5000
    Message:
      type: object
      required: [id, sender, recipient, content, status, sentAt]
      properties:
        id:
          type: string
          format: uuid
        sender:
          $ref: '#/components/schemas/UserSummary'
        recipient:
          $ref: '#/components/schemas/UserSummary'
        content:
          type: string
        status:
          $ref: '#/components/schemas/MessageStatus'
        sentAt:
          type: string
          format: date-time
        deliveredAt:
          type: string
          format: date-time
          nullable: true
        readAt:
          type: string
          format: date-time
          nullable: true
    Notification:
      type: object
      required: [id, recipientId, notificationType, title, body, read, createdAt]
      properties:
        id:
          type: string
          format: uuid
        recipientId:
          type: string
          format: uuid
        actor:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/UserSummary'
        notificationType:
          $ref: '#/components/schemas/NotificationType'
        title:
          type: string
        body:
          type: string
        targetType:
          type: string
          nullable: true
        targetId:
          type: string
          format: uuid
          nullable: true
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
        readAt:
          type: string
          format: date-time
          nullable: true
    NotificationCountResponse:
      type: object
      required: [unreadCount]
      properties:
        unreadCount:
          type: integer
          format: int64
...
    UserSummary:
      type: object
      required: [id, username]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
          format: uri
          nullable: true
...
    MessagePage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        nextCursor:
          type: string
          nullable: true
    NotificationPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
        nextCursor:
          type: string
          nullable: true
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Request validation failed.
        details:
          type: array
          items:
            type: string
...
    PostVisibility:
      type: string
      enum: [PUBLIC, FOLLOWERS_ONLY, PRIVATE]
    PostStatus:
      type: string
      enum: [DRAFT, PUBLISHED, HIDDEN, DELETED]
    MediaType:
      type: string
      enum: [IMAGE, VIDEO, AUDIO, DOCUMENT]
...
    ReactionType:
      type: string
      enum: [LIKE, LOVE, LAUGH, SAD, ANGRY]
    FollowStatus:
      type: string
      enum: [PENDING, ACCEPTED, BLOCKED]
    MessageStatus:
      type: string
      enum: [SENT, DELIVERED, READ, DELETED]
    NotificationType:
      type: string
      enum: [FOLLOW, COMMENT, REACTION, MESSAGE, REPORT_UPDATE, SYSTEM]
...