# Social Media Platform/src/main/java/com/tinsae/socialmediaplatform/profile/controller/ProfileController.java
package com.tinsae.socialmediaplatform.profile.controller;

import com.tinsae.socialmediaplatform.profile.dto.ProfileResponse;
import com.tinsae.socialmediaplatform.profile.dto.UpdateProfileRequest;
import com.tinsae.socialmediaplatform.profile.service.ProfileService;
import jakarta.validation.Valid;
import org.springframework.security.core.annotation.AuthenticationPrincipal;
import org.springframework.security.oauth2.jwt.Jwt;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PatchMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.UUID;

@RestController
@RequestMapping("/api/users/{userId}/profile")
public class ProfileController {

    private final ProfileService profileService;

    public ProfileController(ProfileService profileService) {
        this.profileService = profileService;
    }

    @GetMapping
    public ProfileResponse getProfile(@PathVariable UUID userId) {
        return profileService.getProfile(userId);
    }

    @PatchMapping
    public ProfileResponse updateProfile(
            @PathVariable UUID userId,
            @Valid @RequestBody UpdateProfileRequest request,
            @AuthenticationPrincipal Jwt jwt
    ) {
        return profileService.updateProfile(userId, request, jwt);
    }
}